feat: support multiple projects
