+++ src/types/project.ts
@@ -1,3 +1,5 @@
+import { Risk } from './risk';
+
 export interface ProjectMeta {
   projectName: string;
   projectManager: string;
@@ -5,4 +7,10 @@ export interface ProjectMeta {
   startDate: string; // ISO
   endDate: string; // ISO
   riskPlan: string;
+}
+
+export interface Project {
+  id: string;
+  meta: ProjectMeta;
+  risks: Risk[];
 }
